+++ lib/RN2483-Arduino-Library/src/rn2xx3.h
@@ -210,9 +210,9 @@ public:
 
   uint8_t get_busy_count() const;
 
-  rn2xx3_handler::RN_state wait_command_finished(unsigned long timeout = 10000);
+  rn2xx3_handler::RN_state wait_command_finished(uint32_t timeout = 10000);
 
-  rn2xx3_handler::RN_state wait_command_accepted(unsigned long timeout = 10000);
+  rn2xx3_handler::RN_state wait_command_accepted(uint32_t timeout = 10000);
 
   bool command_finished() const;
 
@@ -266,7 +266,7 @@ public:
    * The RN2xx3 accepts values from 100 to 4294967296.
    * Rumour has it that you need to do a autobaud() after the module wakes up again.
    */
-  void                     sleep(long msec);
+  void                     sleep(int32_t msec);
 
   /*
    * Send a raw command to the RN2xx3 module.
