+++ src/src/DataStructs/TimingStats.h
@@ -204,7 +204,7 @@ extern std::map<int, TimingStats> pluginStats;
 extern std::map<int, TimingStats> controllerStats;
 extern std::map<int, TimingStats> networkStats;
 extern std::map<TimingStatsElements, TimingStats> miscStats;
-extern unsigned long timingstats_last_reset;
+extern uint32_t timingstats_last_reset;
 
 # define START_TIMER const uint32_t statisticsTimerStart(micros());
 # define STOP_TIMER_TASK(T, F) stopTimerTask(T, F, statisticsTimerStart);
