+++ src/src/Helpers/Dallas1WireHelper.h
@@ -239,8 +239,8 @@ void Dallas_write_bit(uint8_t v,
 void Dallas_write_bit_ISR(uint8_t   v,
                           int8_t    gpio_pin_rx,
                           int8_t    gpio_pin_tx,
-                          long      low_time,
-                          long      high_time,
+                          int32_t      low_time,
+                          int32_t      high_time,
                           uint32_t& start);
 
 /*********************************************************************************************\
